Fixes to progress value type and missed out status update.

+++ src/main/java/address/controller/MainController.java
@@ -554,7 +554,7 @@ public class MainController extends UiController{
         return statusBarFooterController.getUpdateStatusReader();
     }
 
-    public Consumer<Long> getUpdateProgressReader() {
+    public Consumer<Double> getUpdateProgressReader() {
         return statusBarFooterController.getUpdateProgressReader();
     }
